+++ package.json
@@ -14,7 +14,8 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "remark": "^12.0.0",
-    "remark-html": "^12.0.0"
+    "remark-html": "^12.0.0",
+    "sass": "^1.32.8"
   },
   "devDependencies": {
     "@types/node": "^14.14.31",
