+++ pages/_app.tsx
@@ -1,4 +1,5 @@
 import { AppProps } from 'next/app'
+import '../styles/global.scss'
 
 function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />
