+++ pages/posts/[id].tsx
@@ -2,7 +2,7 @@ import Layout from '../../components/layout'
 import { getAllPostIds, getPostData } from '../../lib/posts'
 import Head from 'next/head'
 import Date from '../../components/date'
-import utilStyles from '../../styles/utils.module.css'
+import utilStyles from '../../styles/utils.module.scss'
 import { GetStaticProps, GetStaticPaths } from 'next'
 
 export default function Post({
